Further refine the "include `.pids`" hack. (#9773)
### Problem

As described in #9758, there is a workaround in place to allow `pantsd` to watch its pid file in a way that is easily exposed to the Python `pantsd` services. But users who used `pantsd` before we disabled `watchman` will have a socket file at `.pids/watchman/watchman.sock` which needs to continue to be ignored.

### Solution

Refine the workaround to ensure that that subdirectory continues to be hidden. Once less `pantsd` code is on the python side of the FFI boundary we'll have more flexibility to fix this.

### Result

Fixes #9758.

[ci skip-rust-tests]
[ci skip-jvm-tests]