Refactor Workunit types (#9759)
### Problem

In the process of making workunits more sophisticated, it's become clear that the distinction between a `StartedWorkunit` and a `Workunit` no longer makes sense, and leads to a lot of code duplication in trying to handle the mostly-identical `StartedWorkunit` and `WorkUnit` types.

### Solution

This commit unifies workunits into a single type named `Workunit`, and adds a new member `state` of a new type `WorkunitState` to distinguish between started and completed workunits. Also the `WorkUnitStore` type is renamed to `WorkunitStore` to be consistent about the camel-casing of the type.