Remove forced copying in local load_bytes_with (#9793)
### Problem

The LMDB API allows for direct references into its storage, which will always be MMAP'ed. But because it has a blocking API, we interact with it on threads that have been spawned for that purpose.

To allow for interacting with the database in a way that avoids copying data into userspace, the `load_bytes_with` method takes a function that will be called on the blocking thread with a reference directly to that memory. This allows for minimizing copies of data, and avoiding holding full copies of database entries in memory.

But at some point a while back (before async/await made dealing with references easy again), `load_bytes_with` started passing a `Bytes` instance to its callback. And constructing a `Bytes` instance with anything other than a static memory reference (which this isn't) requires copying into the `Bytes` instance to give it ownership. This meant that we weren't actually taking advantage of the odd shape of `load_bytes_with`.

### Solution

Switch the local `load_bytes_with` back to providing a reference into the database's owned memory, and document the reason for its shape. In separate commits, port the `serverset` crate and code that touches it to async/await. Finally, adjust the remote `store_bytes` to accept a reference to avoid potential copies there as well.

### Result

Less memory usage, and less copying of data. In particular: cases that read from the local database in order to copy elsewhere (such as `materialize_directory`, `ensure_remote_has_recursive`, and BRFS `read`) will now copy data directly out of the database and into the destination.

`ensure_remote_has_recursive` should now hold onto only [one chunk's worth](https://github.com/pantsbuild/pants/blob/4f45871814c15c0f41b335fc80f98780b9b38d92/src/python/pants/option/global_options.py#L752-L758) of data at a time per file it is uploading, which might be sufficient to fix #9395 (although the other changes mentioned there will likely still be useful).