Redo approach: use PYTHONPATH instead of init-hook
init-hook doesn't allow us to support any source dependencies, which is a non-starter.

This setup also allows us to auto-configure Pylint to work with the plugin, rather than needing users to add special config.

Finally, this ensures that non-Pylint Pants will understand the target because we make users set up their source roots in a certain way. This means users can write tests for their plugin.

--

This still fails. I think PEX has a special interaction with PYTHONPATH that we need to figure out. But, it's hopefully solvable..
# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]
# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]