Remove stale and private `pants.contrib.buildgen` plugin (#9811)
Turns out, this was only a library and never registered any goals like `buildgen` or any other public API for users. Instead, users typically rely on `fsqio` or the Go implementation of `buildgen.go` to get this functionality.

This library was also never marked "API public".

This code is very out-of-date now, such as using `BuildFile`. Because it's unused and presents a maintenance burden, we remove it here.

[ci skip-rust-tests]
[ci skip-jvm-tests]