Remove unused variants experiment (#9820)
Variants were a way to express in a BUILD file which of several possible values a field should use, e.g. which Thrift config to use. We ended up not using this mechanism.

Here, we still reserve the `@` symbol in addresses for any possible future use, such as reviving this feature.

[ci skip-rust-tests]
[ci skip-jvm-tests]