Render failures via explicit traceback construction rather than graph walks (#9797)
### Problem

In a near-future change, we would like to stop storing failures in the engine's `Graph` in order to improve the experience under `pantsd` (where memoized failures are more problematic than they are otherwise). But `Graph::trace` currently uses the failures stored in the graph to render traces: this "persist failures and then come back later to scrape them" approach is more complicated than it needs to be.

### Solution

Add traceback-like information to a `Failure` by calling a straightforward `Failure::with_pushed_frame` method generically at each `Node` as it fails. The method has one callsite in our `Node` wrapper, which should make it easy to adjust the "frame" information that we'd like to render here: for now, I went with the `user_facing_name` of the `Node` (which means that some Nodes will not be rendered).

### Result

The `Graph::trace` method is deleted, helping to unblock removing failures from the graph. Our rust/python tracebacks also become easier to manipulate in the python-side `Scheduler._raise_on_error` method.

See the changes to the python tests for an example of how the output with `--print-exception-stacktrace` changes.