Fix console UI redraw issues (#9812)
### Problem

With the new console UI, there are a number of console glitches caused by the library that animates the swimlanes not correctly clearing them when some other pants component tries to write to stdout.

### Solution

Whenever we call `ConsoleUI::write_stdout` or `ConsoleUI::with_console_ui_disabled`, we need to de-initialize the screen entirely by calling `teardown`. This drops the data structures that the swimlane-drawing library makes use of causing the screen to be correctly cleared. Doing this in a futures-safe way required some modification to the CFFI functions that call `Scheduler::write_stdout` and `Scheduler::with_console_ui_disabled`.

An earlier version of this code required some additional modifications in `console.py`. We were able to architect this commit in such a way as to make those modifications unnecessary, but in the process we improved the type annotations in `console.py` and those are left in this commit.