Make the `Dependencies` field an AsyncField (#9841)
Soon, we will add two new features:

1) Dependency injection for codegen runtime platforms, e.g. `protobuf_library`'s automatically getting injected with the Python runtime wheel, if configured by the user to do so.
2) Dependency inference, meaning that users can leave off `dependencies` from BUILD files and Pants will figure out the appropriate values from their import statements.

Both of these mechanisms will require the engine to work. So, we can no longer get away with `Dependencies` being a `PrimitiveField` and we instead make it an `AsyncField`.

For now, the rule is extremely simple. The key change is to call sites - this should be the only change necessary for those call sites to automatically work with both features.

### Importance of one rule

Like we do with `sources`, we have only one rule to hydrate dependencies, rather than having a distinct rule for each possible subclass of the `Dependencies` field. 

This is essential to allow us to centralize the implementation of both features to one single place. While we could create helper rules and have us expect subclasses to call those helper rules in their own rules, it would be very error-prone and easy to forgot to do.

Similar to `Sources`, we will add "hooks" to the superclass to allow subclasses to customize functionality. For example, we could add a hook to ban certain addresses. Or, we could add a hook to inject dependencies.

[ci skip-rust-tests]
[ci skip-jvm-tests]