Explicitly handle conflict between --v2-ui and --dynamic-ui (#9849)
### Problem

Now that --v2-ui is deprecated and its value is redirected to that of --dynamic-ui, we are effectively treating this boolean as false by default unless one of those two flags is explicitly set, which we don't want. This seems to be a weird edge case of the `dest` faculty on the option-specification framework.

### Solution

Use the `resolve_conflicting_options` helper function to explicitly handle conflicts between the two flags and remove the `dest`. Also remove the `passive` option since we will never try to use this flag if v2 isn't set. And explicitly set --no-dynamic-ui on the invocation of pants in the bootstrapper so that that invocation of pants doesn't itself spawn a dynamic UI.