Add support for dependency injection and use with V2 Protobuf (#9851)
We want to allow rule authors to create rules that will automatically inject dependencies for certain target types. For example, we want to use this with codegen to automatically inject the runtime libraries needed for the generated code to work properly.

Thanks to the Target API + unions, we can do this easily and also have the following features:

1) A dependency injection rule can ask for any additional params, such as asking for a subsystem.
     * We use this with the protobuf rule to read from a new `--protoc-runtime-targets` option. 
2)  You can have multiple distinct dependency injection rules in use at the same time.
     * For example, we may provide a rule to inject `ProtobufDependencies`. A plugin author could also write a rule to inject all `Dependencies` fields and subclasses. Both will get used together.

### Does not use "precise" dependency injection for codegen

In V2, we have one single `protobuf_library`, rather than `python_protobuf_library` etc. When hydrating `Sources`, we require the call site to disambiguate which of the N languages should be generated by specifying, for example, `PythonSources`.

Here, we do not use this precision. For a `protobuf_library`, we will inject all dependencies specified in `--protoc-runtime-targets`, regardless of the target type.

We do this in part because we cannot know the type of an `Address` until we have converted it into a `Target`. The scope for `resolve_dependencies` is to simply give back the `Addresses` specified in the `dependencies` field, not to filter out irrelevant target types. For example, there is nothing stopping a user from having a `python_library` explicitly depend on a `java_library`.

Instead, that filtering has been happening at the call sites, such as in the file `importable_python_sources.py`. It works well to filter there because the filtering can be arbitrary and complex, such as "get me all targets with PythonBinarySources _and_ EntryPoint". Often, we don't want to filter, like `dependencies.py`.

[ci skip-jvm-tests]
[ci skip-rust-tests]