Add log::Level to WorkunitMetadata (#9850)
This commit is pre-work for a more sophisticated notion of logging in pants that takes into account the current workunit. In order to carry this out, we will need to tag workunits with a log level. This commit adds that field to the `WorkunitMetadata` struct, setting it to `LogLevel::Info` in all cases - eventually this should be configurable.