Compute source roots in a rule.
Previously we computed the source root for a path
outside the engine, via a utillity method. However we
want to support marking source roots using a sentinel
file, and this requires filesystem access in the computation,
meaning that has to be done in a rule.

This change introduces a set of rules for finding the source root
for a path, and converts all v2 code to use those rules.

[ci skip-rust-tests]
[ci skip-jvm-tests]