Fix fmt and lint re-runs under pantsd (#9869)
### Problem

The `FmtFieldSets` and `LinterFieldSets` types extended `Collection` (which implements useful `eq`/`hash`), but then added an additional optional field, `prior_formatter_result`. But that additional field was not included in the `eq`/`hash` for the type, and this meant that instances of these objects would be equal regardless of the previous result. This resulted in `pantsd` memoizing the wrong input for runs: in short, havoc.

### Solution

Extract `StyleRequest` to replace `FmtFieldSets` and `LinterFieldSets`, and fix `__eq__` for that type by extending dataclass. Additionally, align all `Fmt` and `Lint` `FieldSets` types around the name `Request`.

### Result

Operation of `fmt` and `lint` is correct under `pantsd`.

[ci skip-rust-tests]
[ci skip-jvm-tests]