Add support for mutable, global, append only caches (#9852)
### Problem

As described in [Mutable Caching for Processes](https://docs.google.com/document/d/1n_MVVGjrkTKTPKHqRPlyfFzQyx2QioclMG_Q3DMUgYk/edit#) (thank you for the feedback!), we believe that there are two primary usecases for mutable caching in processes (see the document for more detail):
1. global append only
2. local mutable

The first usecase refers to well-behaved systems with their own global append only caches: typically, resolvers like pip/PEX, Coursier, or a conglomeration of other such artifact caches. The second usecase is out of scope for this PR.

### Solution

This change addresses the first usecase by adding support to pants for `append_only_caches` in the `Process` API. It also begins to use such a cache for PEX.

In separate commits, also removes the `unsafe_local_only_files_..` setting to fix #8314 (which the design doc indicates that we will not want to maintain in a pure v2 world), and adds the builder pattern for `Process` for use in tests.

### Result

v2 usages of PEX will concurrently read and write a shared cache in a configurable location. Followup changes can add support for automatically pruning these caches, or propagating the environment variables in remoting via sidechannel information (see the document for more info).