Deprecate `pants.ini` in favor of `pants.toml` (#9879)
We originally did not deprecate `pants.ini` per a request from Twitter, as they have dozens of scripts that read from the INI file and it would have taken substantial time to fix.

However, Twitter will no longer be upgrading to the latest version of Pants. Further, according to users on Slack, the vast majority of users have upgraded to pants.toml without much issue.

We want to get rid of `pants.ini` because it presents some maintenance cost. For example, we're exploring moving part of options parsing to Rust. It will be much easier to do if we only ever have TOML config files, and don't need to support INI.

[ci skip-rust-tests]
[ci skip-jvm-tests]