Rework internal `contrib_plugin` to work with the Target API (#9889)
### Problem

V2 `setup-py` doesn't understand `contrib_plugin` because it relies on mutating the V1 target to have a `provides` field. This doesn't fit in with the V2 target model, where target declarations are no more than 3 lines and should not be mutating their fields.

### Solution

Instead of dynamically adding a `provides` field, have the user explicitly give a `provides` value. Augment the `contrib_setup_py` object so that it can behave the same as before.

Also, remove the unused `PantsPlugin` target type.

### Result

We can now use V2 `setup-py` to build Pants and its plugins.

[ci skip-rust-tests]
[ci skip-jvm-tests]