`BoolField` can  leave off a default to represent tri-states (#9893)
Stu pointed out that it can be valuable to know when a `BoolField` was left undefined. For example, this can be used to know to use a global default value.

Indeed, we had one `python_binary` field that uses this exact pattern and I didn't know how to port. We can now support this use case.

[ci skip-rust-tests]
[ci skip-jvm-tests]