Move remaining Pants wheel building logic to Python (#9898)
We already had most of the logic in Python, but still were running `cp` in Bash. 

This is prework for needing to properly handle when `dist/` could have any of `cp36m`, `cp37m`, or `cp38` wheels in the top-level `dist/` folder. Right now, we assume it will be only one wheel per package.

[ci skip-rust-tests]
[ci skip-jvm-tests]