Build pantsbuild.pants as cp36m, cp37m, and cp38 wheels (rather than abi3) (#9881)
This is necessary for us to move from CFFI to the CPython crate.

Thanks to this change, we can now release a pants.pex that can run with Python 3.6, 3.7, or 3.8, whereas we previously required 3.6.

Note that this means we no longer claim support for running via Python 3.9. It's unlikely Pants worked with 3.9 in the first place, but now we will need to intentionally opt in to it by releasing a new `cp39` wheel.