Log workunit creation and completion (#9894)
### Problem

Now that we have workunits with levels, we want to provide more information about when a workunit starts and ends.

### Solution

Emit a normal log entry when a workunit is created (if the UI is not enabled) and when one completes (always), whose level is the same as that of the associated workunit. This commit also ensured that `goal_rule`s were always created with a description, so that they would always have the `Info` log level be set and be visible with the default pants rule visibility. 

### Result

https://asciinema.org/a/334927 is an asciinema recording of the behavior of a run with the `fmt` rule.