Add a Python import parser for dependency inference (#9900)
At a high level, dependency inference for Python is simply:

1. Parse each source file's AST to extract its import statements.
2. Map those imports to either first-party or third-party targets, if possible.

This implements part 1 by using a mix of regex + AST parsing developed at Toolchain for its buildgen implementation. Notably, this allows us to infer imports from strings like this:

```python
importlib.import_module("example.subdir.Foo")
```

### Works with Python 2

We use `typed-ast` to be able to parse Python 2 files, even though we run Pants with Python 3.

For parsing Python 3 code, we use the stdlib because the Python 3 syntax is generally backward-compatible. The only divergences across versions at the syntactic level are `async`/`await` not being reserved in < 3.7 (deprecated in 3.6), and 3.8 introducing the walrus operator. As explained in a comment, we don't optimize for `async` `await` given the low usage of Python 3.5. There is no solution for 3.8 other than running Pants with Python 3.8, unfortunately.

[ci skip-jvm-tests]