Attach stdout/stderr digests to workunits (#9906)
### Problem

One of the additional pieces of information we would like workunits report is the stdout and stderr of a process, for workunits associated with running a process.

### Solution

This commit represents the first part of the work involved in making this possible, namely, attaching `Digest`s representing the stdout and stderr output of processes to a new `artifacts` key on workunit Python objects. This involves adding new `stdout_digest` and `stderr_digest` fields on the `FallibleProcessResultWithPlatform` type in addition to the extant `stdout` and `stderr` `Bytes` fields. 

Note that this is not enough for the Digests to be useful in Python. A follow-up commit will add new foreign-function APIs that will allow the Python code that uses streaming workunits to materialize these Digests to bytes without using Engine rules directly, and to ensure that a remote store is populated with the contents of these digests. Once that is does, it should also be possible to drop the `stdout` and `stderr` fields from `FallibleProcessResultWithPlatform` completely, since the same information is currently redundantly contained in both fields.

In order to test these effectively, it was also necessary to make a slight change to the `TestBase` unit test framework. Now, a test that subclasses this class can set a class variable called `additional_options` with a list of additional bootstrap options that will be passed to the execution engine for the test. In this case the necessary option was `--no-process-execution-use-local-cache`, which forced a workunit associated with actually running a local `Process` to always be executed.