Log timestamps should be in the local time zone (#9927)
This makes timestamp'd log messages a little bit shorter in the console and friendlier for a human user. If UTC timestamps are desired (for instance in a CI/CD context on a remote machine), the build machine's timezone can be set to UTC. Also use exactly two digits of subsecond precision for log timestamps.