Allow @rules to set their levels, and default to debug (#9940)
### Problem

#9894 and #9921 added workunit completion logging in order to improve the experience when the `--dynamic-ui` is disabled, and to prepare to render streaming warnings/errors and stdio when tests or lints fail, even when the UI is enabled. But further tweaks are needed in order to achieve minimal output in the common case.

### Solution

After those changes, it's become clear that `@rules` should default to debug level, because it continues to be the case that processes are the bread and butter of our runs. Even if/when we implement #7907, the processes that run will continue to represent the bulk of the work and the thing that we cache.

This change defaults all `@rules` to debug, and allows `@rules` to set their own `level`s. Additionally, it removes `@named_rule`, because it's become clear that any `@rule` should be able to set a `canonical_name` (all rules already have one, so we might as well let them set it), `description`, or `level`.

### Result

After this change, only processes (that actually ran, as opposed to hitting cache) will be rendered. `@rules` can opt in to higher log levels, but most should not (unless they are computationally expensive...?): 
```
$ ./v2 --changed-parent=master fmt
18:35:35:904 [INFO] Completed: Run Black on 14 targets: src/python/pants/backend/awslambda/python, src/python/pants/backend/codegen/protobuf/python, src/python/pants/backend/python/lint/bandit, src/python/pants/backend/python/lint/... (403 characters truncated)
18:35:38:233 [INFO] Completed: Run Docformatter on 14 targets: src/python/pants/backend/awslambda/python, src/python/pants/backend/codegen/protobuf/python, src/python/pants/backend/python/lint/bandit, src/python/pants/backend/pytho... (410 characters truncated)
18:35:40:530 [INFO] Completed: Run isort on 14 targets: src/python/pants/backend/awslambda/python, src/python/pants/backend/codegen/protobuf/python, src/python/pants/backend/python/lint/bandit, src/python/pants/backend/python/lint/... (403 characters truncated)
𐄂 Black made changes.
reformatted src/python/pants/engine/internals/engine_test.py
All done! ✨ 🍰 ✨
1 file reformatted, 84 files left unchanged.


✓ Docformatter made no changes.

𐄂 isort made changes.
Fixing src/python/pants/engine/internals/engine_test.py
```