Default to invalidating for `requirements.txt` to ameliorate #7022. (#9946)
### Problem

As explained in #7022, macros that open files can have untracked file dependencies that `pantsd` cannot account for. There is only one (relatively common) case of this: the `python_requirements` macro (which creates a target per line in a `requirements.txt` file).

### Solution

We need to address #7022 in a fundamental way before 2.0. But for now we can ameliorate the only known case by defaulting to invalidating for `requirements.txt` files.

[ci skip-rust-tests]
[ci skip-jvm-tests]