Split client-related logic out of the PantsDaemon class. (#9949)
### Problem

The `PantsDaemon` class was playing double duty as both the entrypoint for the pantsd server, and as a launcher for the client. We will soon need to make significant changes there in order to support #8200 and #7654.

### Solution

Extract a `PantsDaemonProcessManager` base class for process metadata reads/writes and a `PantsDaemonClient` subclass to consume the metadata to decide whether to launch the server. This is 100% code moves... no logic changed.

### Result

The `PantsDaemon` class is now exclusively a server, and `PantsDaemonClient` is now exclusively a client.

[ci skip-rust-tests]
[ci skip-jvm-tests]