Fix lease extension, and make it recursive (#9956)
As was misdiagnosed on #9942, extension of leases is currently broken due to `Store::lease_all` using a separate implementation of `fn lease` that is not aware of `VersionedFingerprint` and thus has a different key. This means that the leases being read in `Store::shrink` are different from the leases being extended in `Store::lease_all`.

Additionally, lease extension is not currently recursive into all values held in memory in the `Graph`, which might mean that a value held in memory in `pantsd` can become invalidated.

Fix the first issue by removing the duplicate implementation of `fn lease`, and making `ShardedLmdb` more directly responsible for lease management: add a test to cover the failing case.

In separate commits, adjust our lease extension of in-memory content to be recursive into `Snapshots` and `ProcessResults`.

Fixes #9942.

[ci skip-jvm-tests]