remoting: add initial unit tests for streaming client and fix a bug (#9966)
### Problem

The streaming remote execution client did not have unit tests when it landed because we wanted to land it early to avoid some merge conflicts. There was also a bug related to decoding an empty string (for the operation name) from the applicable protobuf.

### Solution

1. Adds some initial unit tests.
2. Fixes a bug in decoding empty operation names found with the unit tests.
3. Adds better logging.

### Result

The failing unit test now passes.