Release the gil while polling workunits. (#9969)
### Problem

As reported on #9926 and shown in https://github.com/pantsbuild/pants/files/4732596/gdb.txt, threads 5 and 25 were deadlocked accessing the `WorkunitStore` (in `complete_workunit` and `poll_session_workunits`, respectively) and the GIL (to get `Node` information, and via the `poll_session_workunits` method in `interface.rs`, respectively).

### Solution

Release the GIL before interacting with the `WorkunitStore` in `poll_session_workunits`. I'd like to refactor `interface.rs` to make this kind of issue more challenging to trigger, but this fix is self-contained. 

### Result

Fixes #9926 some more.

[ci skip-jvm-tests]