Create FFI functions for handling digests on workunits (#9965)
### Problem

In order to make practical use of the `Digests` that are attached to Workunits as of https://github.com/pantsbuild/pants/pull/9906 , it is necessary to provide a way to get bytes from a Digest in a way that doesn't use an `@rule`, as well as to ensure that the contents of these digests are uploaded to a remote store.

### Solution

Provide a FFI function `digest_to_bytes` that can extract bytes from a `Digest` with just a reference to the `Scheduler`, bypassing the rest of the engine infrastructure. Also expose the `ensure_remote_has_recursive` store function to Python.