Apply named caches to remote execution platform properties (#10009)
### Problem

We'd like to begin experimenting with mutable caches in remote execution, and received upstream feedback that it would be good to do so via platform properties. Additionally, there was feedback that hardcoding cache locations would make it challenging to use this facility on Windows. 

### Solution

Apply the cache values to the remote execution platform properties. Additionally, apply the upstream feedback: `append_only_caches` is now a dict from cache names to relative paths within the workspace.

### Result

The end-user `Process` API has changed, and we can begin experimenting with mutable caches in remote execution.