Remove `pyopenssl`, `parametrized`, and `wheel` dependencies (#10023)
`pyopenssl` was only being used to pin a transitive dep back in 2017. We don't need to do this anymore because `requests` is more up-to-date. We should also be using a constraints file for this purpose, rather than `requirements.txt`.

`parametrized` is similar to Pytest's parametrization feature. Because we want to encourage tests to use Pytest-style, it's better to use that instead. While Pytest cannot parametrize class-based tests, those can be parametrized through other means like inner helper functions. 

`wheel` was not being used by anything.