Add third party modules to Python dependency inference (#10025)
For this to work, we need a mapping of 3rd party modules to their owning targets. We use a default value of using the same name as the requirement, with some normalization, e.g. `Django` becomes the module `django`. Users can also specify the parameter `modules` to a `python_requirement` or specify `module_mapping` to `python_requirements`. 

If the user has >1 `python_requirement_library` that refers to the same 3rd party requirement, then we do not try to infer the dependency due to ambiguity.