Default to non strict timestamps in zip
When doing our pants bundle, you may encounter issue with
`PANTS_COMPILE_ZINC_INCREMENTAL=true`, because it will create zips with
modification set to 0.

For example, on our project, we sometimes have a cryptic error about
zips and date: 'ZIP does not support timestamps before 1980'

In fact, when digging in code, it is clearly related to
strict_timestamps in python code.

Defaulting it to `False` doesn't seem a bad idea.