Expose new StreamingWorkunitContext on streaming workunits callback (#10034)
### Problem

In order for streaming workunit clients to call the `digests_to_bytes` and `ensure_remote_has_recursive` methods with `Digest` arguments, they need a reference to the scheduler. However, we don't want to just expose a direct reference to the scheduler.

### Solution

Create a new type `StreamingWorkunitContext` that contains a reference to the scheduler, and pass it as an additional kwarg. This type exposes those two methods on the Scheduler, allowing streaming workunit clients to access them.